requirements.txt
google-adk==1.14.0
google-generativeai==0.8.5
pandas==2.3.2
pydantic==2.11.9
python-dotenv==1.1.1
pytest==8.4.1
rouge-score==0.1.2
tabulate==0.9.0

